third_party/WebKit/LayoutTests/fast/workers/worker-constructor-expected.txt

Test Worker constructor functionality. Should print a series of PASS messages, followed with DONE.

PASS: toString exception propagated correctly.
PASS: trying to create workers recursively resulted in an exception (RangeError: Maximum call stack size exceeded)
PASS: invoking Worker constructor without arguments resulted in an exception (TypeError: Failed to construct 'Worker': 1 argument required, but only 0 present.)
PASS: onerror invoked for an empty script URL, resolving to this HTML document's URL.
PASS: invoking Worker constructor with invalid script URL resulted in an exception (SyntaxError: Failed to construct 'Worker': 'http://invalid:123$' is not a valid URL.)
PASS: onerror invoked for a script that could not be loaded.
PASS: Successfully created worker.
DONE
